requirements.txt
Brotli==1.0.9
certifi==2022.9.14
charset-normalizer==2.1.1
click==8.1.3
ConfigArgParse==1.5.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==2.0.2
greenlet==1.1.3
idna==3.4
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.0.1
locust==2.12.1
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.3
psutil==5.9.2
pyzmq==24.0.1
requests==2.28.1
roundrobin==0.0.4
scikit-surprise==1.1.1
scipy==1.5.4
six==1.15.0
typing-extensions==4.3.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
gunicorn==20.1.0
pandas==1.4.4